Move all helper methods to rake_helper.rb
