+++ lib/tasks/build.rb
@@ -1,5 +1,5 @@
-require_relative 'build_utils'
-include BuildUtils
+require_relative 'rake_helper'
+include RakeHelper
 
 namespace :build do
